Backend/requirements.txt
# CUCM Log Collector Backend - Requirements
# Python 3.11+

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Configuration
PyYAML==6.0.2

# SSH Client
asyncssh==2.18.0

# Cryptography (required by asyncssh)
cryptography==44.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1  # For testing FastAPI endpoints

# Optional: Development tools
python-dotenv==1.0.1  # For local .env file support
